src/pf_copilot/requirements.txt
# generative ai SDK dependencies
openai
pydantic<2
--extra-index-url=https://azuremlsdktestpypi.azureedge.net/Create-Dev-Index/102333570
azureml-metrics[prompt-flow]==0.1.0.102333570
--extra-index-url https://azuremlsdktestpypi.azureedge.net/test-azure-ai-generative
azure-ai-generative==0.5.0
azureml-rag[cognitive_search]==0.1.24.1

# promptflow dependencies, these should be optional in the future
# hardcoded the version of azureml-mlflow here for faster Docker image building speed
azureml-mlflow==1.53.0
promptflow[azure]
promptflow-tools
keyring

# langchain dependencies, these should be optional in the future
# langchain moved Embeddings from langchain.embeddings.base to langchain.schema.embeddings while azureml-rag is still referencing it. 
# once azureml-rag is updated to 2.0+, we can remove the langchain reference from this file
langchain<0.0.293

# Project specifi dependency
unstructured
